feat(api): moving endpoints only used by IAM to the internal module
