+++ api/cgrants/api.py
@@ -127,22 +127,17 @@ def _get_contributor_statistics_for_protocol(address: str) -> dict:
     }
 
 
-@api.get(
-    "/contributor_statistics",
-    response=ContributorStatistics,
-    auth=cg_api_key,
-)
-def contributor_statistics(
-    request, address: str | None = None, github_id: str | None = None
-):
-    if not address:
-        return JsonResponse(
-            {
-                "error": "Bad request, 'address' is missing or invalid. A valid address is required."
-            },
-            status=400,
-        )
+# @api.get(
+#    "/contributor_statistics",
+#    response=ContributorStatistics,
+#    auth=cg_api_key,
+# )
+# def contributor_statistics(
+#    request, address: str
+# ):
 
+
+def handle_get_contributor_statistics(address: str):
     if not is_valid_address(address):
         raise InvalidAddressException()
 
@@ -167,115 +162,118 @@ def contributor_statistics(
     return JsonResponse(combined_contributions)
 
 
-@api.get(
-    "/allo/contributor_statistics",
-    response=ContributorStatistics,
-    auth=cg_api_key,
-)
-def allo_contributor_statistics(request, address: str | None = None):
-    if not address:
-        return JsonResponse(
-            {"error": "Bad request, 'address' parameter is missing or invalid"},
-            status=400,
-        )
-
-    if address:
-        address = address.lower()
-
-    response_for_protocol = _get_contributor_statistics_for_protocol(address)
-
-    return JsonResponse(response_for_protocol)
-
-
-def _get_grantee_statistics(identifier, identifier_type):
-    if identifier_type == IdentifierType.HANDLE:
-        grant_identifier_query = Q(admin_profile__handle=identifier)
-
-        contribution_identifier_query = Q(
-            subscription__grant__admin_profile__handle=identifier
-        ) & ~Q(subscription__contributor_profile__handle=identifier)
-    else:
-        grant_identifier_query = Q(admin_profile__github_id=identifier)
-
-        contribution_identifier_query = Q(
-            subscription__grant__admin_profile__github_id=identifier
-        ) & ~Q(subscription__contributor_profile__github_id=identifier)
-
-    # Get number of owned grants
-    num_owned_grants = Grant.objects.filter(
-        grant_identifier_query, hidden=False, active=True, is_clr_eligible=True
-    ).count()
-
-    # Get the total amount of contrinutors for one users grants that where not squelched and are not the owner himself
-    all_squelched = SquelchProfile.objects.filter(active=True).values_list(
-        "profile_id", flat=True
-    )
-
-    num_grant_contributors = (
-        Contribution.objects.filter(
-            contribution_identifier_query,
-            success=True,
-            subscription__is_mainnet=True,
-            subscription__grant__hidden=False,
-            subscription__grant__active=True,
-            subscription__grant__is_clr_eligible=True,
-        )
-        .exclude(subscription__contributor_profile_id__in=all_squelched)
-        .order_by("subscription__contributor_profile_id")
-        .values("subscription__contributor_profile_id")
-        .distinct()
-        .count()
-    )
-
-    # Get the total amount of contributions received by the owned grants (excluding the contributions made by the owner)
-    total_contribution_amount = Contribution.objects.filter(
-        contribution_identifier_query,
-        success=True,
-        subscription__is_mainnet=True,
-        subscription__grant__hidden=False,
-        subscription__grant__active=True,
-        subscription__grant__is_clr_eligible=True,
-    ).aggregate(Sum("amount_per_period_usdt"))["amount_per_period_usdt__sum"]
-    total_contribution_amount = (
-        total_contribution_amount if total_contribution_amount is not None else 0
-    )
-
-    # [IAM] As an IAM server, I want to issue stamps for grant owners whose project have tagged matching-eligibel in an eco-system and/or cause round
-    num_grants_in_eco_and_cause_rounds = Grant.objects.filter(
-        grant_identifier_query,
-        hidden=False,
-        active=True,
-        is_clr_eligible=True,
-        clr_calculations__grantclr__type__in=["ecosystem", "cause"],
-    ).count()
-
-    return JsonResponse(
-        {
-            "num_owned_grants": num_owned_grants,
-            "num_grant_contributors": num_grant_contributors,
-            "num_grants_in_eco_and_cause_rounds": num_grants_in_eco_and_cause_rounds,
-            "total_contribution_amount": total_contribution_amount,
-        }
-    )
+# Unused
+# @api.get(
+#     "/allo/contributor_statistics",
+#     response=ContributorStatistics,
+#     auth=cg_api_key,
+# )
+# def allo_contributor_statistics(request, address: str | None = None):
+#     if not address:
+#         return JsonResponse(
+#             {"error": "Bad request, 'address' parameter is missing or invalid"},
+#             status=400,
+#         )
+#
+#     if address:
+#         address = address.lower()
+#
+#     response_for_protocol = _get_contributor_statistics_for_protocol(address)
+#
+#     return JsonResponse(response_for_protocol)
 
 
-@api.get(
-    "/grantee_statistics",
-    response=GranteeStatistics,
-    auth=cg_api_key,
-)
-def grantee_statistics(
-    request, handle: str | None = None, github_id: str | None = None
-):
-    if not handle and not github_id:
-        return JsonResponse(
-            {
-                "error": "Bad request, 'handle' and 'github_id' parameter is missing or invalid. Either one is required."
-            },
-            status=400,
-        )
+# Unused
+# def _get_grantee_statistics(identifier, identifier_type):
+#    if identifier_type == IdentifierType.HANDLE:
+#        grant_identifier_query = Q(admin_profile__handle=identifier)
+#
+#        contribution_identifier_query = Q(
+#            subscription__grant__admin_profile__handle=identifier
+#        ) & ~Q(subscription__contributor_profile__handle=identifier)
+#    else:
+#        grant_identifier_query = Q(admin_profile__github_id=identifier)
+#
+#        contribution_identifier_query = Q(
+#            subscription__grant__admin_profile__github_id=identifier
+#        ) & ~Q(subscription__contributor_profile__github_id=identifier)
+#
+#    # Get number of owned grants
+#    num_owned_grants = Grant.objects.filter(
+#        grant_identifier_query, hidden=False, active=True, is_clr_eligible=True
+#    ).count()
+#
+#    # Get the total amount of contrinutors for one users grants that where not squelched and are not the owner himself
+#    all_squelched = SquelchProfile.objects.filter(active=True).values_list(
+#        "profile_id", flat=True
+#    )
+#
+#    num_grant_contributors = (
+#        Contribution.objects.filter(
+#            contribution_identifier_query,
+#            success=True,
+#            subscription__is_mainnet=True,
+#            subscription__grant__hidden=False,
+#            subscription__grant__active=True,
+#            subscription__grant__is_clr_eligible=True,
+#        )
+#        .exclude(subscription__contributor_profile_id__in=all_squelched)
+#        .order_by("subscription__contributor_profile_id")
+#        .values("subscription__contributor_profile_id")
+#        .distinct()
+#        .count()
+#    )
+#
+#    # Get the total amount of contributions received by the owned grants (excluding the contributions made by the owner)
+#    total_contribution_amount = Contribution.objects.filter(
+#        contribution_identifier_query,
+#        success=True,
+#        subscription__is_mainnet=True,
+#        subscription__grant__hidden=False,
+#        subscription__grant__active=True,
+#        subscription__grant__is_clr_eligible=True,
+#    ).aggregate(Sum("amount_per_period_usdt"))["amount_per_period_usdt__sum"]
+#    total_contribution_amount = (
+#        total_contribution_amount if total_contribution_amount is not None else 0
+#    )
+#
+#    # [IAM] As an IAM server, I want to issue stamps for grant owners whose project have tagged matching-eligibel in an eco-system and/or cause round
+#    num_grants_in_eco_and_cause_rounds = Grant.objects.filter(
+#        grant_identifier_query,
+#        hidden=False,
+#        active=True,
+#        is_clr_eligible=True,
+#        clr_calculations__grantclr__type__in=["ecosystem", "cause"],
+#    ).count()
+#
+#    return JsonResponse(
+#        {
+#            "num_owned_grants": num_owned_grants,
+#            "num_grant_contributors": num_grant_contributors,
+#            "num_grants_in_eco_and_cause_rounds": num_grants_in_eco_and_cause_rounds,
+#            "total_contribution_amount": total_contribution_amount,
+#        }
+#    )
 
-    if handle:
-        return _get_grantee_statistics(handle, IdentifierType.HANDLE)
-    else:
-        return _get_grantee_statistics(github_id, IdentifierType.GITHUB_ID)
+
+# Unused
+# @api.get(
+#     "/grantee_statistics",
+#     response=GranteeStatistics,
+#     auth=cg_api_key,
+# )
+# def grantee_statistics(
+#     request, handle: str | None = None, github_id: str | None = None
+# ):
+#     if not handle and not github_id:
+#         return JsonResponse(
+#             {
+#                 "error": "Bad request, 'handle' and 'github_id' parameter is missing or invalid. Either one is required."
+#             },
+#             status=400,
+#         )
+#
+#     if handle:
+#         return _get_grantee_statistics(handle, IdentifierType.HANDLE)
+#     else:
+#         return _get_grantee_statistics(github_id, IdentifierType.GITHUB_ID)
