+++ api/stake/api.py
@@ -1,42 +1,42 @@
 from typing import List
 
 from django.db.models import Q
-from ninja_extra import NinjaExtraAPI
 
 import api_logging as logging
-from registry.api.utils import is_valid_address, with_read_db
-from registry.exceptions import InvalidAddressException, StakingRequestError
+from registry.api.utils import with_read_db
+from registry.exceptions import StakingRequestError
 from stake.models import Stake
-from stake.schema import ErrorMessageResponse, GetSchemaResponse, StakeSchema
+from stake.schema import StakeSchema
 from trusta_labs.api import CgrantsApiKey
 
 secret_key = CgrantsApiKey()
 
 log = logging.getLogger(__name__)
 
-api = NinjaExtraAPI(urls_namespace="stake")
+# api = NinjaExtraAPI(urls_namespace="stake")
 
 
-@api.get(
-    "/gtc/{str:address}",
-    auth=secret_key,
-    response={
-        200: GetSchemaResponse,
-        400: ErrorMessageResponse,
-    },
-    summary="Retrieve GTC stake amounts for the GTC Staking stamp",
-    description="Get self and community GTC stakes for an address",
-)
-def get_gtc_stake(request, address: str) -> GetSchemaResponse:
-    """
-    Get relevant GTC stakes for an address
-    """
-    if not is_valid_address(address):
-        raise InvalidAddressException()
+# Currently no public enabled endpoint for this
+# @api.get(
+#     "/gtc/{str:address}",
+#     auth=???,
+#     response={
+#         200: StakeResponse,
+#         400: ErrorMessageResponse,
+#     },
+#     summary="Retrieve GTC stake amounts for the GTC Staking stamp",
+#     description="Get self and community GTC stakes for an address",
+# )
+# def get_gtc_stake(request, address: str) -> StakeResponse:
+#     """
+#     Get relevant GTC stakes for an address
+#     """
+#   if not is_valid_address(address):
+#       raise InvalidAddressException()
 
-    get_stake_response = handle_get_gtc_stake(address)
-    response = GetSchemaResponse(items=get_stake_response)
-    return response
+#   get_stake_response = handle_get_gtc_stake(address)
+#   response = StakeResponse(items=get_stake_response)
+#   return response
 
 
 def handle_get_gtc_stake(address: str) -> List[StakeSchema]:
